Move udp discovery to `NetworkTransport`
