+++ ShortDev.Microsoft.ConnectedDevices/Messages/Discovery/DiscoveryType.cs
@@ -2,6 +2,6 @@
 
 public enum DiscoveryType : byte
 {
-    RresenceRequest,
-    RresenceResponse,
+    PresenceRequest,
+    PresenceResponse,
 }
